chore: reuse search helpers
